[ft #[162233812]] moving db config from server.js to config.js
