+++ app/server.js
@@ -19,19 +19,6 @@ import { ensureAuthenticated } from './middlewares/authentification';
 
 dotenv.config();
 
-// read the virtual environement
-const env = process.env.NODE_ENV;
-
-// get the correponding database URI
-const DATABASEURI = process.env[`PGDATABASE_${env}`];
-// build the config object using the database URI and other env
-const dbConfigObject = {
-  user: process.env.PGUSER,
-  host: process.env.PGHOST,
-  database: DATABASEURI,
-  password: process.env.PGPASSWORD,
-  port: process.env.PGPORT,
-};
 const app = express();
 
 app.get('/', (req, resp) => {
@@ -65,4 +52,5 @@ app.use(error5OOHandler);
 app.get('*', error4O4Handler);
 app.listen(process.env.PORT || 3000);
 
-export { app, dbConfigObject };
+// eslint-disable-next-line import/prefer-default-export
+export { app };
