+++ package.json
@@ -67,7 +67,8 @@
     "pg": "^7.6.1",
     "prettier-eslint": "^8.8.2",
     "request": "^2.88.0",
-    "server-destroy": "^1.0.1"
+    "server-destroy": "^1.0.1",
+    "url": "^0.11.0"
   },
   "nyc": {
     "all": true,
